requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dj-rest-auth==7.0.1
Django==4.2.21
django-allauth==65.10.0
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
Flask==3.1.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
pycparser==2.22
PyJWT==2.10.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
Werkzeug==3.1.3
